Hack to resolve lack of AudioContext support in IE11
IE11 doesn't support WebAudio. MIDI.js in its current form doesn't
support AudioTag anymore. This will be resolved, but in the meantime we
need a fallback for older browsers. This hack just makes sure that the
Pad (the only controller in use right now) doesn't cause MIDI to fall on
its face if WebAudio isn't available, and makes sure we don't try to
create an audio context unless it'll actually succeed.